Improved the code for retrieving data from OpenFoodFacts API. Now it correctly picks every nutritional value as a float, along with the info about if it is solid or liquid, and organic or not. Changed the type value of "energy" from int to float since in some cases, energy is provided as a decimal value.

+++ src/main/java/com/cappellinispirito/ispw_project_202223_jfx/Item.java
@@ -20,7 +20,7 @@ public class Item {
     //construction method
     public Item(
                 //basics for nutriscore
-                int calories,
+                float calories,
                 float sugars,
                 float saturatedFattyAcids,
                 float salt,
